api/requirements.txt
langchain
langchain-core
langchain_community
langchain-experimental
docx2txt
pypdf
python-multipart
streamlit
langchain-groq>=0.1
groq>=0.9
pinecone-client>=4
langchain-pinecone>=0.1
sentence-transformers>=2.7
torch>=2.2
fastapi>=0.111
uvicorn[standard]>=0.30
psycopg2-binary>=2.9
python-dotenv>=1.0
langchain_huggingface
google-generativeai>=0.8.0
transformers>=4.40.0
pymupdf>=1.24.0
pdfplumber>=0.11.0

